Adding semantic checks
The parser validates the mapping semantically by checking that
references to source and target schemas (in `EXISTS` and `WITH` clauses,
respectively) are previously defined (in the `FOREACH` and `WHERE`
clauses). These checks don't consider the validation of the schema
mapping with respect to a metadata catalog, that is, the parser doesn't
compare against a given catalog to see if the referenced relations
actually exist.