Adding generation of QueryGraph
Given a mapping m, a graph is constructed, called the query graph. The
graph includes a node for each target expression that can be
constructed over the variables in the exists clause of the mapping.
Source expressions that provide values for the target will also have
nodes. Furthermore, the graph includes links between nodes to
reflect parent-child relationships and equality predicates,
respectively. This structure can be further refined to support Skolem
function generation; at that point, it contains all the information
needed to generate the query.

This commit doesn't include the refinement of the graph (i.e. doesn't
annotate the nodes)